feat: [MT-74] (WIP) Integrate my store GET & POST API
